Move numpy clone functions from kernel the python core
