+++ Framework/PythonInterface/core/inc/MantidPythonInterface/core/Converters/NumpyFunctions.h
@@ -11,6 +11,7 @@
 #pragma GCC system_header
 #endif
 
+#include "MantidPythonInterface/core/DllConfig.h"
 #include <boost/python/detail/wrap_python.hpp>
 
 // Forward declare the numpy array types
@@ -33,13 +34,16 @@ namespace PythonInterface {
 namespace Converters {
 namespace Impl {
 /// equivalent to macro PyArray_IterNew
-PyObject *func_PyArray_IterNew(PyArrayObject *arr);
+MANTID_PYTHONINTERFACE_CORE_DLL PyObject *
+func_PyArray_IterNew(PyArrayObject *arr);
 /// equivalent to macro PyArray_NewFromDescr
-PyArrayObject *func_PyArray_NewFromDescr(int datatype, const int ndims,
-                                         Py_intptr_t *dims);
-PyArrayObject *func_PyArray_NewFromDescr(const char *datadescr, const int ndims,
-                                         Py_intptr_t *dims);
-PyArray_Descr *func_PyArray_Descr(const char *datadescr);
+MANTID_PYTHONINTERFACE_CORE_DLL PyArrayObject *
+func_PyArray_NewFromDescr(int datatype, const int ndims, Py_intptr_t *dims);
+MANTID_PYTHONINTERFACE_CORE_DLL PyArrayObject *
+func_PyArray_NewFromDescr(const char *datadescr, const int ndims,
+                          Py_intptr_t *dims);
+MANTID_PYTHONINTERFACE_CORE_DLL PyArray_Descr *
+func_PyArray_Descr(const char *datadescr);
 } // namespace Impl
 } // namespace Converters
 } // namespace PythonInterface
